requirements.txt
fastapi
uvicorn[standard]
anthropic
openai
requests
python-dotenv
aiofiles
